Drop lto=true from Cargo.toml
Not forcing LTO is more representative of how the crate would actually be used in third-party code.

It has no meaningful effect on benchmarks for PhastFT, but RustFFT is 15% faster on some sizes without LTO. This gives us a more accurate baseline. It also eases development - with LTO incremental build times are very long.